refactor validate-credentials
